requirements.txt
flask==2.0.2
flask-cors==3.0.10
gunicorn==20.1.0
h5py==3.6.0
numpy==1.19.5
pandas==1.3.4
scikit-learn==1.0.1
google-cloud-logging==2.7.0
google-cloud-storage==1.43.0

